requirements.txt
requests==2.24.0
python-dotenv==0.14.0
shortuuid==1.0.1
python-telegram-bot==12.8
redis==3.5.3
more-itertools==8.5.0
environs==9.3.2
geopy==2.1.0